Close #3299 - Add XVNC port Number in Administration summary page .
Done.

The port can be see in the "Java system properties" button on the mail admin page.

This is very specific and doesn't require a dedicated conditional (Linux +DISPLAY env) IHM in the admin page.

A version will be deployed on http://c8o_dev.convertigo.net as soon as the build is available.